requirements.txt
asgiref==3.8.1
bcrypt==4.3.0
Django==5.2.3
django-cors-headers==4.7.0
djangorestframework==3.16.0
pillow==11.2.1
qrcode==8.2
sqlparse==0.5.3
